[UPDATE] #BTS-130 : 신작 검색 Refactor
