+++ src/main/java/com/readme/sections/repository/NovelCardsRepository.java
@@ -1,6 +1,7 @@
 package com.readme.sections.repository;
 
 import com.readme.sections.model.NovelCards;
+import java.util.Date;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -10,4 +11,5 @@ import org.springframework.stereotype.Repository;
 public interface NovelCardsRepository extends MongoRepository<NovelCards, Long> {
     Page<NovelCards> findAllByTagsNameOrTitleContaining(String keyword, String keyword2, Pageable pageable);
     Page<NovelCards> findAllByGenreAndSerializationStatus(String genre, String serializationStatus, Pageable pageable);
+    Page<NovelCards> findAllByStartDateBetween(Date oneMonthAgo, Date now, Pageable pageable);
 }
